migrate: move banner to components directory
- Migrated banner component files to src/components/banner/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved banners.ts query file to query.ts
- Moved banners.ts formatted type file to formatted-type.ts
- Moved banners.mock.json to mock.json
- Moved banners.test.tsx to query.test.ts
- Moved banners test snapshot to __snapshots__/query.test.ts.snap
- Updated import statements across codebase:
  - src/templates/common/pageLayout/index.tsx updated banner import and BannersData type import
  - src/templates/common/pageLayout/index.test.tsx updated banner import
  - src/templates/common/banners/promoBanner/index.tsx updated type import
  - src/data/queries/index.ts updated query import
- Fixed internal imports:
  - template.test.tsx: ./template instead of ./
  - template.tsx: ./formatted-type instead of @/types/formatted/banners
  - query.ts: ./formatted-type instead of @/types/formatted/banners
  - query.test.ts: ./mock.json and ./query instead of external paths
- Removed empty src/templates/common/banners/banner/ directory
- All TypeScript compilation checks passing
- All tests passing (4 template tests, 5 query tests with 2 snapshots)

This banner component handles VA banner alerts with dismissible options and alert types.
It works with the banners query that fetches banner data from Drupal based on page paths.
The component supports both basic banners and promo banners through shared types and query logic.
