+++ src/components/banner/formatted-type.ts
