+++ src/components/banner/template.tsx
@@ -1,5 +1,5 @@
 import { VaBanner } from '@department-of-veterans-affairs/component-library/dist/react-bindings'
-import { Banner as FormattedBanner } from '@/types/formatted/banners'
+import { Banner as FormattedBanner } from './formatted-type'
 
 export const Banner = ({
   id,
