+++ src/data/queries/index.ts
@@ -5,7 +5,7 @@ import * as AlertBlock from '@/data/queries/alertBlock'
 import * as AlertNonReusable from '@/data/queries/alertNonReusable'
 import * as AlertSingle from '@/data/queries/alertSingle'
 import * as AudienceTopics from '@/data/queries/audienceTopics'
-import * as Banners from '@/data/queries/banners'
+import * as Banners from '@/components/banner/query'
 import * as BenefitsHub from '@/data/queries/benefitsHubLinks'
 import * as Button from '@/data/queries/button'
 import * as CollapsiblePanel from '@/data/queries/collapsiblePanel'
