+++ src/templates/common/banners/promoBanner/index.tsx
@@ -1,5 +1,5 @@
 import { VaPromoBanner } from '@department-of-veterans-affairs/component-library/dist/react-bindings'
-import { PromoBanner as FormattedPromoBanner } from '@/types/formatted/banners'
+import { PromoBanner as FormattedPromoBanner } from '@/components/banner/formatted-type'
 
 export const PromoBanner = ({
   id,
