+++ src/templates/common/pageLayout/index.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from 'test-utils'
 import '@testing-library/jest-dom'
 import { PageLayout, formatBannerType } from './index'
 import { PromoBanner } from '@/templates/common/banners/promoBanner'
-import { Banner } from '@/templates/common/banners/banner'
+import { Banner } from '@/components/banner/template'
 import { BANNER_RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
 
 const children = <div></div>
