+++ src/templates/common/pageLayout/index.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
-import { Banner } from '@/templates/common/banners/banner'
+import { Banner } from '@/components/banner/template'
 import { PromoBanner } from '@/templates/common/banners/promoBanner'
 import { FooterLink } from '@/components/footer/formatted-type'
 import { MegaMenuSection } from '@/components/header/formatted-type'
-import { BannersData } from '@/types/formatted/banners'
+import { BannersData } from '@/components/banner/formatted-type'
 import { BANNER_RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
 import { handleSkipLink } from '@/lib/utils/handleSkipLink'
 import { UnpublishedBanner } from '@/templates/common/preview'
